requirements.txt
uvicorn==0.20.0
pandas==1.5.3
numpy==1.24.2
fastapi==0.91.0
openpyxl==3.1.0
SQLAlchemy==2.0.3
st-annotated-text==3.0.0
pyjsparser==2.7.1
psycopg2==2.9.5
plotly==5.11.0
Js2Py==0.74
matplotlib==3.6.2
mysql==0.0.3
pymysql==1.0.2